New script for publishing design docs and extracting functions.
Added a script for publishing design documents.  It doesn't work,
but I don't know why.  MB-18355 was filed.

Because it didn't work, added another script to extract the map
function so I could paste it in to the UI.  This gets around
the problem that it has all of the whitespace encoding inline.